move shared files to shared feature folder
