+++ frontend/src/modules/alarms/limits/AlarmLimitsPage.tsx
@@ -14,7 +14,7 @@ import { useSelector } from 'react-redux';
 import { VentilationMode } from '../../../store/proto/mcu_pb';
 import { getParametersRequestMode } from '../../../store/controller/selectors';
 import { setActiveRotaryReference } from '../../app/Service';
-import ModeBanner, { BannerType } from '../../displays/ModeBanner';
+import ModeBanner, { BannerType } from '../../modes/ModeBanner';
 import { BPM, PERCENT } from '../../info/units';
 import AlarmLimitsModalPopup from './AlarmLimitsModalPopup';
 import ParamAlarms from './ParamAlarms';
