+++ frontend/src/modules/app/screensaver/ScreensaverPage.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { makeStyles, Theme } from '@material-ui/core/styles';
 import { Grid } from '@material-ui/core';
 import { useSelector } from 'react-redux';
-import ModeBanner, { BannerType } from '../../displays/ModeBanner';
+import ModeBanner, { BannerType } from '../../modes/ModeBanner';
 import ValueJumbotron from './ValueJumbotron';
 import ControlJumbotron from './ControlJumbotron';
 import { PERCENT, BMIN, LMIN } from '../../info/units';
