+++ frontend/src/modules/controllers/index.ts
@@ -12,7 +12,6 @@
  * Re-usable controller components.
  */
 export { default as ValueSlider } from './ValueSlider';
-export { default as Knob } from './Knob';
 export { default as TabPanel } from './TabPanel';
 export { default as TestTool } from './TestTool';
 export { default as ToogleValue } from './ToggleValue';
