+++ frontend/src/modules/logs/LogsPage.tsx
@@ -11,7 +11,7 @@ import { LogEvent, LogEventType } from '../../store/proto/mcu_pb';
 import { getActiveLogEventIds, getNextLogEvents } from '../../store/controller/selectors';
 import { setMultiPopupOpen } from '../app/Service';
 import { AlarmModal } from '../alarms/modal/AlarmModal';
-import ModalPopup from '../modals/ModalPopup';
+import ModalPopup from '../shared/ModalPopup';
 
 import SimpleTable, {
   getComparator,
