+++ frontend/src/modules/modals/index.ts
@@ -5,4 +5,3 @@
  *
  */
 export * from './MultiStepWizard';
-export * from './ModalPopup';
