+++ frontend/src/modules/shared/RotaryEncodeController.tsx
