+++ frontend/src/modules/shared/index.ts
@@ -1,0 +1,2 @@
+export { default as Knob } from './Knob';
+export * from './ModalPopup';
